requirements.txt
requests==2.18.4
python-dotenv==0.8.2